Adding 21C w/ Windows PDF
Adding specific slides for teaching at 21C in Windows.
Really, it's anywhere we teach with Windows and the default
install of Python.  This won't use `python3` or `pip3`.

I'm only doing this because of the horrible failure of trying to
use `c9.io` on the crappy laptops we had to use @ Centennial.